requirements.txt
anthropic
python-dotenv
art
rich
pytest
pytest-json-report